Move DropdownMenu-component to separate file
